Move serialize to utility
